Asset Loading / Unloading Test
Hitbox comment cleaning up
Collision Flags now auto populated with the remaining being set to false if not done manually on hitbox creation.

Added Button in mainmenu to move to maingamescene
Added Button in maingamescene to move to mainmenu
Both with some base logging use, to be expanded on in new branch

The system of doing this is unrefined, with naming and procedural oddities which should probably be redone in a later push.